refactor(netlify): extract shared origin validation (#78)
record-resonance and get-resonance each carried their own copy of the
allowed-origins list, the URL.origin same-origin check, and the
boilerplate to produce a 403 when the Origin header was rejected.
During #76 the same deploy-preview origin fix had to be applied to
both files independently.

Pull all of that into netlify/functions/_lib/origin.ts and have both
functions call checkOrigin(request, CORS_HEADERS) — returns a 403
Response if the request should be blocked, null otherwise. The
underscored _lib subdirectory is ignored by Netlify's function
discovery, so the shared module isn't deployed as a function.

getEnv was duplicated for the same reasons (and is needed by the
origin module to look up the production URL env var), so move it
into netlify/functions/_lib/env.ts as well.

Behavior is unchanged: same allowed-origins list, same same-origin
fallback, same FORBIDDEN error body. Missing Origin headers continue
to be allowed through as non-browser callers.

Closes #78.
