+++ netlify/functions/record-resonance.ts
@@ -1,19 +1,8 @@
-declare const Netlify: {
-  env: { get(key: string): string | undefined };
-};
+import { getEnv } from './_lib/env';
+import { checkOrigin } from './_lib/origin';
 
 interface NetlifyHandlerContext {
   deploy?: { context?: string };
-}
-
-function getEnv(key: string): string | undefined {
-  try {
-    const value = Netlify.env.get(key);
-    if (value !== undefined) return value;
-  } catch {
-    // Netlify global not available in local dev
-  }
-  return process.env[key];
 }
 
 interface ResonanceBody {
@@ -84,33 +73,6 @@ function evictStale(): void {
       requestLog.set(key, valid);
     }
   }
-}
-
-const ALLOWED_ORIGINS: string[] = [
-  'http://localhost:4321',
-  'http://localhost:8888',
-];
-
-function getAllowedOrigins(): string[] {
-  const origins = [...ALLOWED_ORIGINS];
-  const siteUrl = getEnv('URL');
-  if (siteUrl) {
-    origins.push(siteUrl);
-  }
-  return origins;
-}
-
-function isAllowedOrigin(origin: string, requestUrl: string): boolean {
-  if (getAllowedOrigins().includes(origin)) return true;
-
-  // Allow same-origin requests (covers production, deploy previews, branch deploys)
-  try {
-    if (origin === new URL(requestUrl).origin) return true;
-  } catch {
-    // malformed URL — fall through
-  }
-
-  return false;
 }
 
 const CORS_HEADERS: Record<string, string> = {
@@ -339,11 +301,8 @@ export default async (request: Request, context: NetlifyHandlerContext) => {
     return errorResponse('Method not allowed', 'METHOD_NOT_ALLOWED', 405);
   }
 
-  // Origin check: block requests from unknown origins, allow missing Origin (non-browser)
-  const origin = request.headers.get('Origin');
-  if (origin && !isAllowedOrigin(origin, request.url)) {
-    return errorResponse('Forbidden', 'FORBIDDEN', 403);
-  }
+  const forbidden = checkOrigin(request, CORS_HEADERS);
+  if (forbidden) return forbidden;
 
   if (!getEnv('GITHUB_TOKEN')) {
     return errorResponse('GitHub integration not configured', 'SERVICE_UNAVAILABLE', 503);
